refactor(ui): split each context handle from its provider component
A provider module that also exports its consumer hook can never be
hot-replaced: React Refresh needs every export to be a component, and the
hook is not one. Editing a provider therefore forced a full reload and lost
the workbench's window layout, open chat, and scroll position — the exact
state a fast-refresh edit loop exists to preserve.

`context/` already had the answer twice over: `showPageDrag.ts` +
`ShowPageDragProvider.tsx`, and `unsavedChangesContext.ts` +
`useUnsavedChanges.ts` + `UnsavedChangesProvider.tsx`. Seven more contexts
now follow it — the handle, the value type and the consumer hooks in a `.ts`
module, the provider component alone in a `.tsx`.

Consumers are untouched: `ToastContext.tsx` became `ToastContext.ts`, so
all 55 `useToast` call sites keep importing from the same specifier. Only
the provider mount points move, which is one file each (App.tsx, AppShell,
ChatPage) plus three tests that wrap a subject in ToastProvider.

The split also gives each mechanism a home. The status polling cadences
belong to the poller, the toast dedupe window and the on-screen row shape
belong to the toast stack, and the cascade offsets and desktop-viewport
check belong to the window manager — all now sit beside the provider that
reads them, while the handle module carries only the contract.

The two media viewers in `ui/` get the same treatment under that directory's
kebab-case convention (`file-viewer-context.ts`, `image-viewer-context.ts`).
Both hooks intentionally return null with no provider mounted, which is a
documented fallback for FileCard and the shared Markdown renderer; splitting
the handle out means those two can ask for a preview without pulling in the
lazy modal that answers it.

Errors: 253 -> 243. react-refresh: 14 -> 4.
