+++ ui/src/components/ui/file-viewer.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
+import { FileViewerContext, type FilePreviewTarget } from './file-viewer-context';
+
 // Session-scoped in-app file preview. ChatPage mounts the provider; a FileCard
 // with a previewable proxy file or a classified local rich file opens the modal. The
 // heavy modal (Shiki, the JSON viewer, the CSV parser) is ``React.lazy``-loaded,
@@ -7,25 +9,6 @@ import * as React from 'react';
 // preview opens. Mirrors ImageViewerProvider so the two media viewers stay
 // structurally consistent; FileCard falls back to a plain link where no
 // provider is mounted (``useFileViewer()`` returns null).
-
-export type FilePreviewTarget =
-  | { kind?: 'media'; url: string; name?: string }
-  | {
-      kind: 'local';
-      path: string;
-      name: string;
-      size: number | null;
-      mime: string | null;
-      ext: string | null;
-    };
-
-type FileViewerContextValue = { open: (target: FilePreviewTarget) => void };
-
-const FileViewerContext = React.createContext<FileViewerContextValue | null>(null);
-
-export function useFileViewer(): FileViewerContextValue | null {
-  return React.useContext(FileViewerContext);
-}
 
 const FileViewerModal = React.lazy(() => import('@/components/ui/file-viewer-modal'));
