+++ ui/src/components/workbench/ChatPage.tsx
@@ -67,7 +67,8 @@ import { Button } from '../ui/button';
 import { ChatImage } from '../ui/chat-image';
 import { FileCard } from '../ui/file-card';
 import { ImageViewerProvider } from '../ui/image-viewer';
-import { FileViewerProvider, useFileViewer } from '../ui/file-viewer';
+import { FileViewerProvider } from '../ui/file-viewer';
+import { useFileViewer } from '../ui/file-viewer-context';
 import { Input } from '../ui/input';
 import { Markdown } from '../ui/markdown';
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover';
